Big refactoring:
Extracting calculating of cards to its own component - which values can be calculated using memoization (by extracting out all the individual keys).

Deciding if should animate based on if the window size is changed or if the matrix xy changed.

Deciding if should remove cards if the filterable cards changed.